+++ mlxtend/evaluate/tests/test_feature_importance.py
@@ -12,6 +12,8 @@ from sklearn.datasets import make_regression
 from sklearn.model_selection import train_test_split
 from sklearn.svm import SVC
 from sklearn.svm import SVR
+from sklearn.metrics import r2_score
+from sklearn.metrics import mean_squared_error
 from mlxtend.utils import assert_raises
 from mlxtend.evaluate import feature_importance_permutation
 
@@ -111,6 +113,66 @@ def test_regression():
     assert imp_vals[0] > 0.2
     assert imp_vals[1] > 0.2
     assert sum(imp_vals[3:]) <= 0.01
+
+
+def test_regression_custom_r2():
+
+    X, y = make_regression(n_samples=1000,
+                           n_features=5,
+                           n_informative=2,
+                           n_targets=1,
+                           random_state=123,
+                           shuffle=False)
+
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.3, random_state=123)
+
+    svm = SVR(kernel='rbf', gamma='auto')
+    svm.fit(X_train, y_train)
+
+    imp_vals, imp_all = feature_importance_permutation(
+        predict_method=svm.predict,
+        X=X_test,
+        y=y_test,
+        metric=r2_score,
+        num_rounds=1,
+        seed=123)
+
+    assert imp_vals.shape == (X_train.shape[1], )
+    assert imp_all.shape == (X_train.shape[1], 1)
+    assert imp_vals[0] > 0.2
+    assert imp_vals[1] > 0.2
+    assert sum(imp_vals[3:]) <= 0.01
+
+
+def test_regression_custom_mse():
+
+    X, y = make_regression(n_samples=1000,
+                           n_features=5,
+                           n_informative=2,
+                           n_targets=1,
+                           random_state=123,
+                           shuffle=False)
+
+    X_train, X_test, y_train, y_test = train_test_split(
+        X, y, test_size=0.3, random_state=123)
+
+    svm = SVR(kernel='rbf', gamma='auto')
+    svm.fit(X_train, y_train)
+
+    imp_vals, imp_all = feature_importance_permutation(
+        predict_method=svm.predict,
+        X=X_test,
+        y=y_test,
+        metric=mean_squared_error,
+        num_rounds=1,
+        seed=123)
+
+    norm_imp_vals = imp_vals / np.abs(imp_vals).max()
+
+    assert imp_vals.shape == (X_train.shape[1], )
+    assert imp_all.shape == (X_train.shape[1], 1)
+    assert norm_imp_vals[0] == -1.
 
 
 def test_n_rounds():